MCP-GateWay/internal/policy/ml/server/requirements.txt
# ML 서버용 의존성
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf>=4.21.6,<5.0dev
torch==2.0.1
transformers==4.30.2
numpy==1.24.3
scikit-learn==1.3.0
safetensors==0.3.1
